+++ src/database/queries/bulkUploadTasks/selectWithPagination.sql
@@ -1,18 +1,18 @@
-SELECT bulk_upload_to_json(bulk_uploads.*) as "object"
-FROM bulk_uploads
+SELECT bulk_upload_task_to_json(bulk_upload_tasks.*) as "object"
+FROM bulk_upload_tasks
 WHERE
   CASE
     WHEN :createdBy::UUID IS NULL THEN
       true
     ELSE
-      bulk_uploads.created_by = :createdBy
+      bulk_upload_tasks.created_by = :createdBy
     END
   AND CASE
     WHEN :authContextKeycloakUserId::UUID IS NULL THEN
       true
     ELSE
       (
-        bulk_uploads.created_by = :authContextKeycloakUserId
+        bulk_upload_tasks.created_by = :authContextKeycloakUserId
         OR :authContextIsAdministrator::boolean
       )
     END