+++ src/database/queries/bulkUploads/selectById.sql
@@ -1,3 +1,0 @@
-SELECT bulk_upload_to_json(bulk_uploads.*) as "object"
-FROM bulk_uploads
-WHERE id = :id;