+++ src/routers/index.ts
@@ -1,7 +1,6 @@
 import express from 'express';
 import { applicationFormsRouter } from './applicationFormsRouter';
 import { baseFieldsRouter } from './baseFieldsRouter';
-import { bulkUploadsRouter } from './bulkUploadsRouter';
 import { dataProvidersRouter } from './dataProviderRouter';
 import { fundersRouter } from './fundersRouter';
 import { opportunitiesRouter } from './opportunitiesRouter';
@@ -12,6 +11,7 @@ import { presignedPostRequestsRouter } from './presignedPostRequestsRouter';
 import { proposalsRouter } from './proposalsRouter';
 import { proposalVersionsRouter } from './proposalVersionsRouter';
 import { sourcesRouter } from './sourcesRouter';
+import { tasksRouter } from './tasksRouter';
 import { usersRouter } from './usersRouter';
 import { documentationRouter } from './documentationRouter';
 
@@ -19,7 +19,7 @@ const rootRouter = express.Router();
 
 rootRouter.use('/applicationForms', applicationFormsRouter);
 rootRouter.use('/baseFields', baseFieldsRouter);
-rootRouter.use('/bulkUploads', bulkUploadsRouter);
+rootRouter.use('/tasks', tasksRouter);
 rootRouter.use('/changemakers', changemakersRouter);
 rootRouter.use('/changemakerProposals', changemakerProposalsRouter);
 rootRouter.use('/dataProviders', dataProvidersRouter);