+++ src/routers/tasksRouter.ts
@@ -1,0 +1,19 @@
+import express from 'express';
+import { bulkUploadTasksHandlers } from '../handlers/bulkUploadTasksHandlers';
+import { requireAuthentication } from '../middleware';
+
+const tasksRouter = express.Router();
+
+tasksRouter.post(
+	'/bulkUploads',
+	requireAuthentication,
+	bulkUploadTasksHandlers.postBulkUploadTask,
+);
+
+tasksRouter.get(
+	'/bulkUploads',
+	requireAuthentication,
+	bulkUploadTasksHandlers.getBulkUploadTasks,
+);
+
+export { tasksRouter };